I FIXED THE WINDOW STACKING I JUST HAD TO SIMPLIFY THE CALCULATION
house height is walls multiplied by floors. each floor is as tall as  the wall height. i just had to multiply the floor number by the wall to get the current floor. everything was already centered omg I CANT BELIEVE THIS TOOK ME A FREAKING MONTH TO FIGURE OUTTTT UHHGHHHH

anyway i also moved the window GRP return to the outside of the for loop for the floor iteration so that it would stop ending the floor creation at the first iteration